activator-seed-template: Update to Finatra RC1 release along with finagle-httpx
Problem
The template is behind in releases of Finatra and Finagle

Solution
Update the template to the latest Finatra release built against the latest Finagle release (which includes the
switch to finagle-httpx from finagle-http). The lastest Finatra release also includes a way to default the
http port programmatically so that we can set the port in code to not conflict with the Activator UI.

Result
An updated template that uses the latest code and doesn't require a user to start the Activator UI on a
different (other than default) port of :8888.